refactor: change misspeled names of components and move person table to separate component
