--- src/components/PeopleLink.tsx
@@ -1,14 +1,0 @@
-import { NavLink } from 'react-router-dom';
-import { Person } from '../types';
-import classNames from 'classnames';
-
-export const PeopleLink = ({ person }: { person: Person }) => {
-  return (
-    <NavLink
-      to={`/people/${person.slug}`}
-      className={classNames({ 'has-text-danger': person.sex === 'f' })}
-    >
-      {person.name}
-    </NavLink>
-  );
-};
